requirements.txt
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cobble==0.1.4
cryptography==46.0.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
google_search_results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.2
mdurl==0.1.2
numpy==2.4.1
openai==2.16.0
openpyxl==3.1.5
pandas==3.0.0
pathvalidate==3.3.1
pdfminer.six==20260107
pillow==12.1.0
prompt_toolkit==3.0.52
puremagic==1.30
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
regex==2026.1.15
requests==2.32.5
rich==14.3.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.5.0
xlsxwriter==3.2.9
youtube-transcript-api==1.2.3
